requirements.txt
fastapi==0.112.2
fastapi-cli==0.0.5
pandas==2.2.2
python-multipart==0.0.9
pydantic-settings==2.4.0